Implemented eslint and addressed errors.
NB errors relate to use of jQuery so rules have been disabled to allow this. The one other error related to the overall function not being used; this is because the function is invoked via a click listener within the html. The rule has therefore been disabled for this line.